requirements.txt
langgraph
fastapi
langserve[all]
uvicorn
groq
langchain
langchain-community
python-dotenv
langchain-groq
python-multipart
PyPDF2
nltk
jinja2
sentence-transformers
langchain-pinecone
pinecone-notebooks
langchain-huggingface
semantic-split
fastembed
langchain_experimental